chore: update returned values from processFrame

+++ libDF/src/android.rs
@@ -215,7 +215,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
         Ok(state) => state,
         Err(err) => {
             log_error(&err);
-            return -10f;
+            return -10 as jfloat;
         }
     };
 
@@ -227,7 +227,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
 
     if !is_direct {
         log_error("ByteBuffer must be direct");
-        return -20f; // Error code for non-direct buffer
+        return -20 as jfloat; // Error code for non-direct buffer
     }
 
     let byte_buffer = unsafe { JByteBuffer::from_raw(byte_buffer.into_raw()) };
@@ -236,7 +236,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
         Ok(capacity) => capacity,
         Err(e) => {
             log_error(&format!("Failed to get direct buffer capacity: {:?}", e));
-            return -30f;
+            return -30 as jfloat;
         }
     };
 
@@ -246,7 +246,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
             state.window_size(),
             buffer_capacity
         ));
-        return -40f;
+        return -40 as jfloat;
     }
 
     // Get buffer pointer and capacity
@@ -254,7 +254,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
         Ok(ptr) => ptr as *mut i16,
         Err(e) => {
             log_error(&format!("Failed to get direct buffer address: {:?}", e));
-            return -50f;
+            return -50 as jfloat;
         }
     };
 
@@ -278,7 +278,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
         Ok(lsnr) => lsnr,
         Err(e) => {
             log_error(&format!("Failed to process audio frame: {:?}", e));
-            return -60f;
+            return -60 as jfloat;
         }
     };
 
@@ -287,7 +287,7 @@ pub extern "C" fn Java_com_rikorose_deepfilternet_NativeDeepFilterNet_processFra
         Some(slice) => slice,
         None => {
             log_error("Failed to get output as slice");
-            return -70f;
+            return -70 as jfloat;
         }
     };
